+++ docs/Changelog.md
@@ -1,6 +1,19 @@
 # Changelog
 
-## 1.12.0 (TBA)
+## 2.0.0 (Feb 2025)
+
+- SDK simplification and added `OIDC` activation feature [(#182)](https://github.com/wultra/mtoken-sdk-android/pull/182)
+    - [Migration guide](Migration-2.0.md)
+
+## 1.12.0 (Oct, 2024)
+
+- Upgraded to `PowerAuthSDK` `1.9.x` [(#165)](https://github.com/wultra/mtoken-sdk-ios/pull/165)
+- PowerAuth "server stack" `1.9+` is now required
+- Document fixes and improvements
+- Added status to `UserOperation` and removed redundant `OperationHistoryEntry` [(#171)](https://github.com/wultra/mtoken-sdk-ios/pull/171)
+- Removed deprecated `IOperationsService` methods [(#171)](https://github.com/wultra/mtoken-sdk-ios/pull/171)
+
+## 1.11.1 (Aug, 2024)
 
 - Added resultTexts to UserOperation [(#152)](https://github.com/wultra/mtoken-sdk-ios/pull/152)
 - Extended PushParser to support parsing of inbox notifications [(#150)](https://github.com/wultra/mtoken-sdk-android/pull/150)